move types.py into utils, move GenesisDataTypedDict to types.py
