--- geth/types.py
@@ -1,7 +1,0 @@
-from typing import (
-    IO,
-    Any,
-    Union,
-)
-
-IO_Any = Union[IO[Any], int, None]
